+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2023-11-06
+
+- Fixed invocations missing new splat
+- Fixed issue if `CreateHardLink` fails due to `terraform` binary in use
+
 ## [2.0.0] - 2023-11-05
 
 - Update config search behavior